+++ client/src/lib/actions/ems-fd.ts
@@ -24,6 +24,7 @@ interface IDispatch {
   medicalRecords?: MedicalRecord[];
   status?: string;
   status2?: string;
+  activeDeputy?: Deputy;
 }
 
 export const createEmsFdDeputy = (data: object) => async (dispatch: Dispatch<IDispatch>) => {
@@ -71,6 +72,7 @@ export const getCurrentEmsStatus = () => async (dispatch: Dispatch<IDispatch>) =
         type: GET_CURRENT_EMS_STATUS,
         status: res.data.deputy?.status || "off-duty",
         status2: res.data.deputy?.status2 || "-",
+        activeDeputy: res.data.deputy || null,
       });
     }
   } catch (e) {